ci-cd: add ci-cd:common generator
