+++ generators/ci-cd/generators/common/templates/docker-registry.yml.ejs
